Adding one parameter to deployment parameters file
In attempting to build the environment using the supplied azuredeploy.paramenters.json I noticed the InstallADFS sub-deployment was failing, this was due to the $cert variable in InstallADFS.ps1 being blank, as the domainFQDN eventually passed as a parameter was not present. Adding it here along with a cautionary note to avoid placing the wildcard, as installADFS.ps1 does this alrady.